gfx/layers/apz/test/mochitest/test_group_bug1534549.html

<!DOCTYPE HTML>
<html>
<head>
    <meta charset="utf-8">
    <title>Tests for bug 1534549</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <script type="application/javascript" src="apz_test_native_event_utils.js"></script>
    <script type="application/javascript" src="apz_test_utils.js"></script>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
    <script type="application/javascript">
    var touch_action_prefs = getPrefs("TOUCH_ACTION");

    var subtests = [
        // Tests that z-index ordering is respected by hit-test info.
        { "file": "helper_touch_action_ordering_zindex.html", "prefs": touch_action_prefs },
        // Tests that complex block/inline background ordering is respected by hit-test info.
        { "file": "helper_touch_action_ordering_block.html", "prefs": touch_action_prefs },
    ];

    if (isApzEnabled()) {
        ok(window.TouchEvent, "Check if TouchEvent is supported (it should be, the test harness forces it on everywhere)");
        if (getPlatform() == "android") {
            // This has a lot of subtests, and Android emulators are slow.
            SimpleTest.requestLongerTimeout(2);
        }

        SimpleTest.waitForExplicitFinish();
        window.onload = function() {
            runSubtestsSeriallyInFreshWindows(subtests)
                .then(SimpleTest.finish, SimpleTest.finishWithFailure);
        };
    }
    </script>
</head>
<body>
</body>
</html>
